Put random loadout in the menu and update styles
