--- src/app/loadout/random/random-loadout.scss
@@ -1,12 +1,0 @@
-.random {
-  position: fixed;
-  backface-visibility: hidden;
-  bottom: 0;
-  bottom: constant(safe-area-inset-bottom);
-  bottom: env(safe-area-inset-bottom);
-  right: 0;
-  padding: 0.5rem;
-  color: black;
-  text-decoration: none;
-  font-weight: normal;
-}
