+++ src/locale/dim.json
@@ -164,7 +164,7 @@
     "Desc": "DIM is moving Engram and Glimmer items from {{store}} to the vault and leaving one space open per item type to prevent anything from going to the Postmaster.",
     "Desc_female": "DIM is moving Engram and Glimmer items from {{store}} to the vault and leaving one space open per item type to prevent anything from going to the Postmaster.",
     "Desc_male": "DIM is moving Engram and Glimmer items from {{store}} to the vault and leaving one space open per item type to prevent anything from going to the Postmaster.",
-    "FarmingMode": "Farming Mode (move items)",
+    "FarmingMode": "Farming Mode (make space)",
     "MakeRoom": {
       "Desc": "DIM is moving only Engram and Glimmer items from {{store}} to the vault or other characters to prevent anything from going to the Postmaster.",
       "Desc_female": "DIM is moving only Engram and Glimmer items from {{store}} to the vault or other characters to prevent anything from going to the Postmaster.",
@@ -470,7 +470,7 @@
     "BurnTypeArc": "Arc",
     "BurnTypeSolar": "Solar",
     "BurnTypeVoid": "Void",
-    "CreateLoadout": "Create Loadout",
+    "CreateLoadout": "New Loadout",
     "EquipItems": "Equip on {{name}}",
     "GeneratedBuilds": "Builds",
     "Loading": "Generating builds... {{loading}}%",
@@ -561,7 +561,10 @@
     "PullFromPostmasterPopupTitle": "Pull from Postmaster",
     "PullMakeSpace": "Space",
     "Random": "Random",
-    "Randomize": "Randomize your equipped weapons, armor, ghost, and artifact?",
+    "Randomize": "Randomize",
+    "WeaponsOnly": "Weapons",
+    "RandomizePrompt": "Randomize your equipped class, weapons, armor and ghost?",
+    "RandomizeWeapons": "Randomize your equipped class and weapons?",
     "RestoreAllItems": "All Items",
     "Save": "Save",
     "SaveAsNew": "Save As New",
